+++ include/hit.hpp
@@ -15,7 +15,7 @@ public:
         t = 1e38;
     }
 
-    Hit(float _t, Material *m, const Vector3f &n) {
+    Hit(double _t, Material *m, const Vector3f &n) {
         t = _t;
         material = m;
         normal = n;
@@ -30,7 +30,7 @@ public:
     // destructor
     ~Hit() = default;
 
-    float getT() const {
+    double getT() const {
         return t;
     }
 
@@ -42,14 +42,14 @@ public:
         return normal;
     }
 
-    void set(float _t, Material *m, const Vector3f &n) {
+    void set(double _t, Material *m, const Vector3f &n) {
         t = _t;
         material = m;
         normal = n;
     }
 
 private:
-    float t;
+    double t;
     Material *material;
     Vector3f normal;
