+++ include/object3d.hpp
@@ -16,7 +16,7 @@ class Object3D {
 
     // Intersect Ray with this object. If hit, store information in hit
     // structure.
-    virtual bool intersect(const Ray &r, Hit &h, float tmin) = 0;
+    virtual bool intersect(const Ray &r, Hit &h, double tmin) = 0;
 
    protected:
     Material *material;
