+++ include/scene_parser.hpp
@@ -70,7 +70,7 @@ class SceneParser {
 
     Vector3f readVector3f();
 
-    float readFloat();
+    double readFloat();
     int readInt();
 
     FILE *file;
