+++ src/mesh.cpp
@@ -7,7 +7,7 @@
 #include <sstream>
 #include <utility>
 
-bool Mesh::intersect(const Ray &r, Hit &h, float tmin) {
+bool Mesh::intersect(const Ray &r, Hit &h, double tmin) {
     // Optional: Change this brute force method into a faster one.
     bool result = false;
     for (int triId = 0; triId < (int)t.size(); ++triId) {
